Made two important additions to the Simulation Library:
(1) Added a new method, rdModel* LoadModel(const string &aModelLibraryName), for loading and creating a model by sending the method the name of the model DLL.  This is significant because now executables can be run for any model simply by providing the name of the model; one does not have to modify the main routine for a specific model and then recompile the main.  This will greatly facilitate scripting and running workflows with different models.  The code for the LoadModel() method is in the file LoadModel.h and LoadModel.cpp.

(2) Added some print methods for printing information out about a model.  The new methods are printBasicInfo() and printDetailedInfo().



git-svn-id: https://simtk.org/svn/opensim/Trunk@1388 551e70df-b0f6-0310-80ee-9b1508f7454c