Major overhaul to SdfastEngine.  It's now compiled into a dll which does not include any of the sdfast functions.  The sdfast model is compiled into its own dll.  Then at run time when the model is loaded and the SdfastEngine is initialized, it gets function pointers to the sdfast functions in the model's dll.
A <model_library> property is added to the SdfastEngine object to specify which library the engine should load.

The specific tools which might encounter sdfast models manually load the SdfastEngine library (using LoadOpenSimLibrary) so that the proper types are registered and the osim file's SdfastEngine section can be properly parsed.

This should help the GUI side because now an sdfast model can be loaded by just specifying the osim file, and it is the SdfastEngine's responsibility to locate and load the model-specific dll.

This still does not resolve the issue of multiple sdfast models loaded at the same time -- it won't work because the sdfast data is static in the library and Windows will not load two instances of the same library (unless they have different names on disk).

git-svn-id: https://simtk.org/svn/opensim/Trunk@2840 551e70df-b0f6-0310-80ee-9b1508f7454c