Added code for the muscle analysis.  Most standard quantities for muscles are now printed out, with the exception of moment arms.  The most desirable data structure for moment arms is different that for the scalars like (tendon length, fiber length, force, etc), so I'll do a separate moment arm analysis.
One important change I did make in SimmDarrylMuscle.cpp is that I multiplied the _activeForce by the activation in computeActuation().  Otherwise it was being set to values corresponding to full activation.

Another interesting thing to note is that the active force can be much greater than what is seen in the tendon (even accouting for pennation angle).  The active force in SimmDarrylMuscle is used to compute the muscle state derivative.  Because of internal damping, this is not the force seen at the tendon.

git-svn-id: https://simtk.org/svn/opensim/Trunk@2905 551e70df-b0f6-0310-80ee-9b1508f7454c