Often when doing simulations, it can be difficult to choose initial states.  A good example is choosing values for the initial lengths of muscles.  For a particular activation level, you have to root solve to find values of fiber lengths that are reasonable (e.g., in quasi-static equilibrium).  Pete added code to the Zajac muscle model and Darryl's muscle model that computes equilibrium values for muscle fiber lengths, but there was not a general way of calling these at the beginning of CMC or a forward simulation, for example.  Different types of actuators may have different states to solve for.
So, I added new methods to the Model, ActuatorSet, and AbstractActuator class called computeEquilibriumForAuxiliaryStates()
(or something similar).  This is just a general hook for computing equilibrium values for states other than the generalized coordinates
and speeds.  Each actuator is responsible for implementing its own version of this method.  The muscle classes, for example, call Pete's
computeIsometricForce() method, which internally also computes a new value for the muscle fiber length.

I have called these new methods only in the ForwardTool and tested the functionality using Kate's upper extremity model (we had a problem of
computing reasonable starting values for the muscle fiber lengths) and on the Gait2354 example.  There is a new flag in AbstractTool called
_solveForEquilibriumForAuxiliaryStates (really long I know-- if you think of a better name let me know), which is false by default, that
can be used to control whether equilibrium states are solved for.


git-svn-id: https://simtk.org/svn/opensim/Trunk@2984 551e70df-b0f6-0310-80ee-9b1508f7454c