More cleanup. ModelComponentSets for Joints and Coordinates are NOT setup by model, but rather the owning Body is solely responsible for setting up its Joint and it in turn must setup its Coordinates. Flat lists of Joints and Coordinates at the model level are populated at Model::setup after BodySet is setup. This avoids duplicate calls to setup for Joints and triplicate calls to setup for Coordinates!
Model default constructor createGroundBodyIfNecessary while convenience (filename) and copy constructor invoke setup().  Adding components also invokes setup for the set to which the component is added too. API users should no longer have to call setup() on model themselves.

CorrectionController no longer blindly adds actuators to the model, it checks first that the actuator does not already exist in the model. Controller maintains a list of actuators it controls but it is no longer assumes ownership of those actuators.








git-svn-id: https://simtk.org/svn/opensim/Trunk@5658 551e70df-b0f6-0310-80ee-9b1508f7454c