Reorganization of Muscle info structures for containing and calculating muscle quantities efficiently. Better documentation of structure quantities by Matt Millard. MuscleLengthInfo is a Position stage cache, FiberVelocityInfo is Velocity stage, and MuscleDynamicsInfo is dependent on Dynamics. This ensures that fiber velocity can always be obtain prior to Dynamics and thus be used by Muscle controllers (e.g. stretch reflex).
Thelen2003Muscle was minimally updated to operate with the new Muscle interface. It appears to be generating the same results. Differences in results seem to be related to more integration steps due to the update to Simbody 3.0 than due to Muscle behavior.

Clarification/simplification of the ModelComponent interface. Made adding of continuous state and discrete variables a singular operation (one state at a time) rather than a vector of states. Also, added facility to add state variable with specified dependent stage. The default remains Dynamics, which the underlying subssytem allocateZ is currently tied to.

The following tests FAILED:
1>  	  5 - testNCSpline (Failed)
1>  	 22 - testVersioning (Failed)
1>  	 43 - testCMC (Failed)

testCMC is failing at running the running simulation and requires further investigation.

git-svn-id: https://simtk.org/svn/opensim/Trunk@6991 551e70df-b0f6-0310-80ee-9b1508f7454c