An exception in calcMuscleDynamicsInfo was being raised when d/dt(KE+PE-W) > tol, tol ~= 0. This exception is handy during a simulation, but is not useful during many other tasks. This exception has been commented out since its impossible to detect what task the muscle is being used for.
The quantity d/dt(KE+PE-W) is being stored in the 0th element of the vector 'userDefinedDynamicsInfo' in the struct MuscleDynamicsInfo, and thus can still be used by us for debugging purposes. 

This is a fragile place to put the value, as some api user in the future might decide to put something else in this spot. If we find that this quantity is useful, then we should consider giving it it's own variable in MuscleDynamicsInfo

git-svn-id: https://simtk.org/svn/opensim/Trunk@7101 551e70df-b0f6-0310-80ee-9b1508f7454c