Lowered default activation from 1.0 to 0.05 for ActivationFiberLengthMuscle. It used to be 0, but that is because old muscle models were changing activations to some min value under the covers. This was calling testInitState to fail. Commented testInitState so that the test is decipherable.
Prescribed motion was not being satisfied in the state returned by initSystem. The assembly analysis now has a check for whether coordinates are locked or prescribed and not just if there are constraints in the constraint set. testAssemblySolver is passing.  This also allowed testMuscles to compute the correct initial system kinetic energy.

testMuscles was modified so that a copy of the muscle is used in the harness. Otherwise, repeated calls to the main simulation driver results in accumulated path points and the path length computed is wrong.
Only testThelen2003Muscle is failing. System energy minus work is off by the potential of the muscle.





git-svn-id: https://simtk.org/svn/opensim/Trunk@7186 551e70df-b0f6-0310-80ee-9b1508f7454c