1. First time commit of the new equilibrium muscle model: Millard2012EquilibriumMuscle. This muscle model has had the activation model changed, and its dynamic equations modified so that when simulated the integrated state will never wander into a region that causes a singularity. This model is being checked in a little early to give the GUI team time to integrate this in with their code. There are a number of to-do's still related to this model:
a. Automatic test code: This muscle model has been given an initial hand test (it conserves energy, its initialization is perfect, and the plots look as I would expect them). This model still needs to be integrated with 'testMuscles' so that these tests are done automatically

b. The pennation compression spring profile might need editing. It causes a pretty rapid change in fiber velocity when it engages at the moment ... it might be more desirable to have it gradually compress.

c. A fitting routine for the ForceVelocityInverseCurve needs to be added

d. A fitting routine for the ActiveForceLengthCurve still needs to be added.

e. Discuss coupling the lower bound of the ActiveForceLengthCurve and the engagement length of the FiberCompressiveForceLengthCurve.

2. RegisterTypes_osimActuators has been modified to include this new muscle model.

3. The FiberCompressiveForceCosPennationCurve has been altered so that this compressive element engages at an angle of 85 degrees, so that this element does not influence the simulation results until a singularity (when pennation = 90 degrees) is imminent.

4. A figure fig_Millard2012EquilibriumMuscle.png has been added to support the doxygen for the new muscle model.

git-svn-id: https://simtk.org/svn/opensim/Trunk@7211 551e70df-b0f6-0310-80ee-9b1508f7454c