Restoration of testStaticOptimization. Test was failing because initialization of updated Thelen2003Muscle was throwing an exception due to 0 activation values being fed to the StaticOptimization analysis.
The AnalyzeTool was fixed so that it does not assume 0 as a state value when none is provided. Instead it relies on the state generated by initSystem (initState), which initializes state variable values according defaults (properties) associated with the model/component.

StaticOptimzation was updated such that equilibrium is re-established based-on the default activation of the muscle. The default is then updated after a new set of activations are computed, such that at each instant the defualt activation corresponds to the previous time. This should improve performance and is less likely to have convergence issues.

testStaticOptimization explicitly uses Object::renameType to test the Thelen2003Muscle_Deprecated. The test using the updated Thelen2003Muscle is temporarily on hold until we assess the new results. The case passes to reflect its agreement with previous results with the deprecated muscle model.

A quick check of testForward by renaming to Thelen2003Muscle_Deprecated also now passes.



git-svn-id: https://simtk.org/svn/opensim/Trunk@7226 551e70df-b0f6-0310-80ee-9b1508f7454c