1. MuscleFirstOrderActivationDynamicModel: The parameter 'minimumActivation' is no longer serializable as per Sherm & Ajay's wise suggestion. This parameter is now a private data member, has a default value of 0.01, and is settable but only through the api.
2. MuscleFixedWidthPennationModel
Fixed a poor (ok really poor!) math choice in calcMuscleLength that Ajay pointed out.

3. Thelen2003Muscle, Millard2012EquilibriumMuscle, Millard2012AccelerationMuscle
a. Updated to work with the changes in MuscleFirstOrderActivationDynamicModel
b. Updated the initialization tolerance to be

tol = 1e-8*maximumIsometricForce();

Unless tol < SimTK::Eps*1000, in which case it is SimTK::Eps*1000. Here tol is the desired tolerance of the initial solution in Newtons.

No new tests failed (and thanks to Ajay now CMC passes! Go Ajay!). Additionally I ran my own main() benchmark on each of the three muscles for a pennated muscle, and for a non-pennated muscle. The results looked just fine (good initial conditions, the derivative of system energy less work was 0 to tolerance).


git-svn-id: https://simtk.org/svn/opensim/Trunk@7276 551e70df-b0f6-0310-80ee-9b1508f7454c