Added Exception handling to MuscleAnalysis and AnalyzeTool to deal with the richer set of Exceptions provided by the new muscle models. This should enable the plotter to function as before without performing any error handling itself.
MuscleAnalysis's record method was separated into three distinct stages:  reporting lengths, reporting forces, and reporting velocities and powers. The latter requiring the evaluation of muscle dynamics to get fiber velocity from an equilibrium model. Each category has its own try-catch block then lets the recording resume even for muscles that fail (throw an Exception). The message of the first offending Exception is outputted and subsequent exceptions at that stage are ignored.

AnalyzeTool's equilibrateMuscles was also put in a try-catch so that analyses may resume, even if one muscle were to fail to equilibrate.  To make this work, Model's equilibrateMuscles traps Exceptions from equilibrate of each individual muscle, but resumes to evaluate each muscle, then finally it throws an Exception if any of the muscles failed to equilibrate and passes along the message of the first Exception to be thrown.

Muscle models were updated so that if they fail to computeInitialFiberEquilibrium(), then they set the fiber length and equilibrium muscle force to be NaN. This enables the analysis to report results for all muscles, up to and including the regions were some muscles fail, thereby providing potentially useful information as to why they failed (can see fiber-length approaching zero, for example, and enabling the analysis to report values for unaffected muscles.

Finally, the allocation and updating of storages was overhauled.  Storages are allocated only once, right in the begin method instead of in the constructors. This no longer requires having Storages for each coordinate in anticipation of which might be selected by the user and eliminates the need for an updateStorages method due to changes to the desired output lists via access methods. This "just-in-time" creation means the list of storages is now only the size of what is needed and we can cycle through the list of storages if for loop, instead of having to hand-pick storages for printing results (to avoid the bulk of storages allocated for unused coordinates for both moments and moment arms).



git-svn-id: https://simtk.org/svn/opensim/Trunk@7309 551e70df-b0f6-0310-80ee-9b1508f7454c