Once these changes have been looked at by a few people, it would be good to see the `Loose Ends' section. I'm game to take care of these loose ends, but not until other eyeballs have seen this, and testMuscles.
1. testMuscles: major overhaul 

a. New correctness test added that asserts that d/dt(KE+PE-W) = 0 to numerical tolerance (1e-8 right now). The old version of this test, which did not include terms internal to the muscle models, remains but is not being used.

b. The new muscle models (Millard2012EquilibriumMuscle, and Millard2012AccelerationMuscle) have been added to testMuscles.

c. The assertion that the test must run in real time or faster has been removed because its not reasonable for this test (3.3% tendon). The Thelen2003Muscle, and Millard2012EquilibriumMuscle models barely run faster than real time in release. The Millard2012AccelerationMuscle runs slower than real time. The speedup factor (simulation time / clock time) is reported, but not used to throw an exception.

d. testPathActuator has been removed from this test because it is not a muscle, and also because I cannot instrument the energy/power transfers internal to this code, so I cannot actually test if this component is conservative. If this test should be used, then it should be in 'testActuator', as it is an actuator and not a muscle.

e. testThelen2003Muscle_Deprecated, testThelen2003MuscleV1, testShutte1993Muscle, and testDelp1990Muscle have been removed because these models either do not implement the new Muscle interface, and so I cannot perform a correctness test on them as I cannot get at the internal energy/power transfers of these models.

2. Muscle:

a. the field 'fiberPower' was partitioned into 'fiberActivePower' and 'fiberPassivePower'. Fiber active power is the part of the fiber power that is associated with muscle activation. Fiber passive power are all of the remaining power terms (associated with the parallel element, and if they exist, the compressive elements and dampers).

b. Added 'getFiberActivePower' and 'getFiberPassivePower' methods

c. Removed the field 'fiberPowerAlongTendon' as it is superfluous for current pennation models that do no work.

3. MuscleAnalysis

a. Removed 'fiberPower' and added 'fiberActivePower' and 'fiberPassivePower'.

4. RigidTendonMuscle, Thelen2003Muscle, Millard2012EquilibriumMuscle, and Millard2012AccelerationMuscle

a. Updated calcMuscleDynamicsInfo to reflect the changes to the struct described in item 2.

5. osimActuators : added the RigidTendonMuscle, Millard2012EquilibriumMuscle and Millard2012AccelerationMuscle. These should have been included in osimActuators (I believe) but have been overlooked.

Loose ends:
1. Remove testThelen2003MuscleV1! It should only be available as a plugin.

2. Add an initialization test for the muscles. Here is a simple candidate test:
a. Fixed length, fixed activation.
b. Initialize the muscle
c. Simulate the muscle - its length should not change during the simulation.

git-svn-id: https://simtk.org/svn/opensim/Trunk@7335 551e70df-b0f6-0310-80ee-9b1508f7454c