I re-ran all of the tests up to testCMC and they all passed. I wasn't willing to wait the 20+ minutes (in release!!!) at 11:06pm for testCMC to complete.
1. MuscleFirstOrderActivationDynamicModel: 

Will now clamp both excitation and activation to be within a legal range when .calcDerivative is called rather than throwing an exception. This will allow a simulation to run with invalid excitation and activation values, which we presume a user will see in the simulation results during their due diligence check of the results. This change was requested by several expert users who where finding the exceptions particularly onerous.

2. testMuscleFirstOrderActivationDynamicModel:

Added a test case to test the clamping.

3. MuscleFixedWidthPennationModel

Provides a functionality to clamp the fiber length, and pennation angle to the shortest numerically stable values (hueristically determined for now). Expert users found the original behaviour of throwing an exception upon receiving an illegal value to make this model too onerous to use.

4. testMuscleFixedWidthPennationModel

Added a test case to ensure that the clamping functionality is working

5. Thelen2003Muscle

a. Put the initialization routine in a try-catch statement, so if it bails (which it can) I can initialize the muscle crappy but legal initial conditions.

b. Clamped the fiberLengths and activations used by the model internally to legal values.

c. Implemented the model switching condition to smoothly switch from Thelen's dynamic equations to a fixed fiber length just prior to the fiber hitting its minimum value and going singular. I spent a great deal of time determining what the appropriate switching condition is to ensure that the resulting fiber kinematic curves for length and pennation are C0 continuous, and fiber velocity and pennation angular velocity curves are transient free across the switch. 

6. testMuscles

a. Did a great deal of house keeping to the main test routine to make it easier to read, and also to use. 

b. Added flags and a spot to test the quality of the model's initial conditions, and also to compare the analytically computed muscle stiffness to a numerical one. These tests have not yet been implemented.

To Do: update the documentation .. sigh


git-svn-id: https://simtk.org/svn/opensim/Trunk@7385 551e70df-b0f6-0310-80ee-9b1508f7454c