Adding states variables to a ModelComponent had a side-effect that their derivatives in computeStateVariableDerivatives had to be in the alphabetical order of the variable names. Although this was consistent with the order of getStateVariableNames(), it was a bug since no component was checking this order. For the most part this went unnoticed because only a few components add state variables and out of sheer luck, activation, fiber-length and fiber-velocity are alphabetically ordered. The MuscleExample, however, was wrong and was detected by Friedl when she tried to extend the behavior of a muscle and was verified by Matt M.
The fix was simply to associate the order of addition/creation of the state variable in the underlying map. 
Now the name, underlying SimTK::Index and creation order of the state variable are tied together. Lists (array, vector, etc..) of state variable names or their values are all assembled/interpreted in this order. 

The MuscleExample standard results were updated to reflect the correction.

All tests are passing.  

git-svn-id: https://simtk.org/svn/opensim/Trunk@7448 551e70df-b0f6-0310-80ee-9b1508f7454c