Fixed some nasty subtleties in OpenSim::Model::initializeState() (this was a previous TODO - involved more work than originally thought - but I believe it is for the better).
Previously, a reference to the model's default state was returned and used as the model's working state. This is not ideal because the default state was then being corrupted which meant that anyone wanting to perform multiple simulations using the same model would have to re-build the model for each simulation. This bug went undetected because no-one was ever performing multiple simulations within a single main() by re-initializing the state using initializeState(). The fix from this checkin involved the following:

1) Created an internal member variable in model.h: SimTK::State _workingState;
2) Rather than returning a reference to updMultibodySystem().updDefaultState(), the _workingState now holds a copy of the default state: getMultibodySystem().getDefaultState() and a writable reference to this will be returned. Therefore, the model's default state is preserved and untouched.
3) Added a convenience method getWorkingState() that returns a reference to _workingState.
4) Wrote some doxygen explaining the above changed in model.h.

This allows the user to get a guaranteed "fresh" state if multiple simulations are to be performed. This is a subtle fix, but important when executing multiple simulations on a given model. Code reviewed with Ajay. Have also had discussions with Sherm about the change.


However, when running test cases, there were a whole bunch that failed. This is a result of several analyses/components that directly accessed the model's default state (assuming that it is the working state). Not good. The solution was to change the references in these analyses/tools.

FROM _model->updMultibodySystem.updDefaultState();
TO _model->getWorkingState();

All the OpenSim tests now pass with this change.

In general, it is not very good practice to assume that the model's default state is the working state because this is part of the Model and therefore should be read only. When performing an analysis/simulation, we should always work on a COPY of the default state so we don't corrupt the default state (otherwise, if we ever need it again, it'll be damaged). Now, there is no need to copy the state after initializeState() because the OpenSim::Model now holds an internal "copy". So you can always return a reference and this will return the reference to the internal copy: i.e. State&s = initializeState().

git-svn-id: https://simtk.org/svn/opensim/Trunk@7567 551e70df-b0f6-0310-80ee-9b1508f7454c