This check-in addresses a bug reported by Matt DeMers regarding the accuracy at which constraints are being satisfied. We originally had the assembly problem setup as best fit to the constraints and any other goals (e.g. matching user-specified coordinate values). When the specified coordinate value(s) are in conflict with the constraint(s) this can lead to an awkward compromise solution where neither the constraint nor the coordinate values are satisfied very well. Therefore, we have opted to strictly enforce the constraint instead. Then if and only if the assembly fails, does it relax the constraint to get a best fit solution. TODO: In this case we should report to the user that the constraint could not be satisfied.
An earlier check-in addressed the issue of ignoring some coordinates, specifically those with the is_free_to_satisfy_constraints set to true. Instead of changing the weight of the coordinate goal to zero, I am now excluding the coordinate from having an associated goal in the first place.

A stripped down version of Matt's model is included as an additional case in testAssemblySolver.cpp.  The test moves the knee angle through its range of motion (as a slider would do in the GUI) and the model is assembled to satisfy the constraint, which requires the knee angle to deviate from the original set value.
We must verify that the slider and value presented in the GUI indicate the assembled configuration and not what the user was trying to set.

The CustomActuatorExample/testCustomActuatorExample.cpp example was updated because the model now starts in a slightly different (less than 1deg) position and this leads to trajectories that differ slightly from the standard. The tolerances for the test were increased for deviations in coordinates speeds and spring force.


git-svn-id: https://simtk.org/svn/opensim/Trunk@7924 551e70df-b0f6-0310-80ee-9b1508f7454c