Fixed a bug in StaticOptimization that was estimating target accelerations from double differentiated coordinates (q) after the model had been assembled. This led to very small non-smooth transitions in q that manifested as spikes in the acceleration by double differentiating the spline and spikes in muscle activations and forces. Since velocities are also adjusted by the assembly to satisfy constraints, the velocity errors are likely to be small and also less sensitive error since they could be differentiated once to estimate the target acceleration. This change was made and acceleration estimates are smooth. This fixes a bug discovered by Matt Demers when running SO on his model/data from 3.0 in 3.1b1. which causes spikes in the muscle force trajectories.
TestStaticOptimization was completely neutered by Matt Millard, in the sense that the test no longer performed the regression to verify that the answers do not change. It was simply testing that SO runs with the different muscle models. I restored it so that it performs the regression test with the Thelen2003Muscle_Deprecated muscle at least. Once we have a MuscleRedundancySolver we can perform unit tests as suggested by Matt M. and restore the regression testing for the different Muscle models.

git-svn-id: https://simtk.org/svn/opensim/Trunk@8107 551e70df-b0f6-0310-80ee-9b1508f7454c