Cleaned up the MuscleMetabolicPowerProbeUmberger2010 header so that the implementation of the internal MuscleMetabolicPowerProbeUmberger2010::MetabolicMuscleParameter comes at the end of the file.
Put all nested class implementations in the .cpp

Making the nested class private to MuscleMetabolicPowerProbeUmberger2010 causes testProbes to fail so that change was backed off.

I also removed use of the MetabolicMuscleParameterSet since it had no members of implementation other than overriding setAuthors. Since properties are just as happy with Set<MetabolicMuscleParameter> so there is one less nested class. testProbes was modified to use Set<MetabolicMuscleParameter>.

As it stands testProbes is not using the probe interface and it is still relying on instantiating the MetabolicMuscleParameter (should be called MetabolicMuscleParameters) and populating using its public accessors. This should no longer be necessary since the Probe was revamped by Tim. Once Tim updates the test and it is free of MetabolicMuscleParameter uses, the class can be made private to MuscleMetabolicPowerProbeUmberger2010.

I also suggest we consider renaming these Probes so that they end with the word Probe that is:
Umberger2010MuscleMetabolicPowerProbe.





git-svn-id: https://simtk.org/svn/opensim/Trunk@8116 551e70df-b0f6-0310-80ee-9b1508f7454c