1) Reverted MetabolicMuscleParameterSet from last checkin (Set<MetabolicMuscleParameter>) to fix xml parsing problems. 2) Moved all method definitions to cpp file (for Bhargava probe (Ajay did this previously for the Umberger probe). 3) Put MetabolicMuscleParameter and MetabolicMuscleParameterSet at the end of the file (for both probes).
Having some issues making MetabolicMuscleParameter and MetabolicMuscleParameterSet private so am leaving it public for now. 

Apart from that, I am now freezing the metabolic probe implementation (pending any major bugs). Sorry it took so long, but this will be a key feature of the 3.1 release so we want to have a good solid interface for the Darpa teams to use.

git-svn-id: https://simtk.org/svn/opensim/Trunk@8118 551e70df-b0f6-0310-80ee-9b1508f7454c